refactor: moved services to CostAccounting.Services
